Update dev doc examples with level prop
