+++ packages/terra-framework-docs/src/terra-dev-site/doc/slide-panel-manager/example/SlidePanelManager.module.scss
@@ -12,4 +12,9 @@
     background-color: #add8e6;
     padding: 1rem;
   }
+
+  .form-select {
+    margin-left: 1rem;
+    max-width: 300px;
+  }
 }
