--- packages/terra-framework-docs/src/terra-dev-site/test/slide-panel-manager/SlidePanelManager.module.scss
@@ -1,20 +1,0 @@
-:local {
-  .example-wrapper {
-    border: 1px solid #d3d3d3;
-    height: 35rem;
-  }
-
-  .content-wrapper {
-    padding: 0.7rem;
-  }
-
-  .disclosure-accessory {
-    background-color: #add8e6;
-    padding: 1rem;
-  }
-
-  .form-select {
-    margin-left: 1rem;
-    max-width: 300px;
-  }
-}
